fixed DeleteOverlay
